Fixed the bugs

+++ controllers/Product.controller.js
@@ -2,11 +2,13 @@ import Product from "../models/ProductModel.js";
 
 export async function getAllProducts(req, res) {
   try {
-    const products = await Product.find({});
+    const products = await Product.find().populate("createdBy", "username");
     res.json(products);
   } catch (error) {
     console.error("Error fetching products:", error);
-    res.status(500).send("Server error");
+    res.status(500).json({
+      error: "Internal server error",
+    });
   }
 }
 
@@ -14,14 +16,22 @@ export async function getProductById(req, res) {
   try {
     const productId = req.params.id;
 
-    const product = await Product.findById(productId);
+    const product = await Product.findById(productId).populate(
+      "createdBy",
+      "username"
+    );
     if (!product) {
-      return res.status(404).send("Product not found");
+      return res.status(404).json({
+        success: false,
+        error: "Product not found",
+      });
     }
     res.json({ success: true, data: product });
   } catch (error) {
     console.error("Error fetching product by id:", error);
-    res.status(500).send("Server error");
+    res.status(500).json({
+      error: "Internal server error",
+    });
   }
 }
 
@@ -32,7 +42,13 @@ export async function addProduct(req, res) {
       return res.status(400).send("All fields are required");
     }
 
-    const newProduct = new Product(product);
+    const newProduct = new Product({
+      name: product.name,
+      price: product.price,
+      descr: product.descr,
+      img: product.img,
+      createdBy: req.user._id, // Assuming req.user is set by authentication middleware
+    });
 
     await newProduct.save();
 
@@ -42,7 +58,9 @@ export async function addProduct(req, res) {
     });
   } catch (error) {
     console.error("Error adding product:", error);
-    res.status(500).send("Server error");
+    res.status(500).json({
+      error: "Internal server error",
+    });
   }
 }
 
@@ -51,14 +69,18 @@ export async function updateProduct(req, res) {
     const productId = req.params.id;
     const product = req.body;
 
-    if (!product.name || !product.price || !product.descr || !product.img) {
-      return res.status(400).send("All fields are required");
-    }
+    const productUser = await Product.findById(productId);
 
     const updatedProduct = await Product.findByIdAndUpdate(productId, product, {
       new: true,
-      runValidators: true,
     });
+
+    if (productUser.createdBy.toString() !== req.user._id.toString()) {
+      return res.status(403).json({
+        success: false,
+        error: "You are not authorized to update this product",
+      });
+    }
 
     res.status(200).json({
       success: true,
@@ -66,23 +88,40 @@ export async function updateProduct(req, res) {
     });
   } catch (error) {
     console.error("Error updating product:", error);
-    res.status(500).send("Server error");
+    res.status(500).json({
+      error: "Internal server error",
+    });
   }
 }
 
 export async function deleteProduct(req, res) {
   try {
     const productId = req.params.id;
-    const deletedProduct = await Product.findByIdAndDelete(productId);
+
+    const deletedProduct = await Product.findById(productId);
     if (!deletedProduct) {
-      return res.status(404).send("Product not found");
+      return res.status(404).json({
+        success: false,
+        error: "Product not found",
+      });
     }
+
+    if (deletedProduct.createdBy.toString() !== req.user._id.toString()) {
+      return res.status(403).json({
+        success: false,
+        error: "You are not authorized to delete this product",
+      });
+    }
+
+    await Product.findByIdAndDelete(productId);
 
     res
       .status(200)
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
     console.error("Error deleting product:", error);
-    res.status(500).send("Server error");
+    res.status(500).json({
+      error: "Internal server error",
+    });
   }
 }
